LayoutTests/storage/indexeddb/setversion-not-blocked-expected.txt

Test that setVersion is not blocked if handle closed in versionchange handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-not-blocked.html"
indexedDB.deleteDatabase(dbname)
blockedEventFired = false
versionchangeEventFired = false
indexedDB.open(dbname)

prepareDatabase():
h1 = event.target.result
h1.onversionchange = h1OnVersionChange
indexedDB.open(dbname, 2)

h1OnVersionChange():
versionchangeEventFired = true
PASS event.oldVersion is 1
PASS event.newVersion is 2
h1.close()

h2OpenBlocked():
PASS event.oldVersion is 1
PASS event.newVersion is 2
blockedEventFired = true

h2UpgradeNeeded():
PASS event.oldVersion is 1
PASS event.newVersion is 2

h2OpenSuccess():
PASS versionchangeEventFired is true
FIXME: blocked should not have fired since connection closed; http://webkit.org/b/71130
FAIL blockedEventFired should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
